chore: change to typescript
