+++ manifest.json
@@ -1,29 +1,29 @@
 {
-	"format_version": 2,
-	"metadata": {
-		"authors": ["HirziDevs"],
-		"url": "https://github.com/HirziDevs",
-		"license": "MIT"
-	},
-	"header": {
-		"name": "§eCCBG §d[1.2.2]§f",
-		"description": "Customize your Cobblestone and Basalt Generator!\nAddons By: §bHirziDevs§f",
-		"uuid": "15c4dbf1-0871-42e5-8b04-114daa5b7ab4",
-		"version": [1, 2, 2],
-		"min_engine_version": [1, 21, 0]
-	},
-	"modules": [
-		{
-			"type": "script",
-			"uuid": "8df2bb30-ab8b-4797-bc58-c589b64fd5f7",
-			"version": [1, 0, 0],
-			"entry": "scripts/main.js"
-		}
-	],
-	"dependencies": [
-		{
-			"module_name": "@minecraft/server",
-			"version": "1.12.0-beta"
-		}
-	]
+  "format_version": 2,
+  "metadata": {
+    "authors": ["HirziDevs", "Hexaa Scoooolzs"],
+    "url": "https://github.com/HirziDevs",
+    "license": "MIT"
+  },
+  "header": {
+    "name": "§eCCBG §d[1.2.2]§f",
+    "description": "Customize your Cobblestone and Basalt Generator!\nAddons By: §bHirziDevs§f\nContributor: §bHexaa Scoooolzs§f",
+    "uuid": "15c4dbf1-0871-42e5-8b04-114daa5b7ab4",
+    "version": [1, 2, 2],
+    "min_engine_version": [1, 21, 0]
+  },
+  "modules": [
+    {
+      "type": "script",
+      "uuid": "8df2bb30-ab8b-4797-bc58-c589b64fd5f7",
+      "version": [1, 0, 0],
+      "entry": "scripts/dist/main.js"
+    }
+  ],
+  "dependencies": [
+    {
+      "module_name": "@minecraft/server",
+      "version": "1.12.0-beta"
+    }
+  ]
 }
