+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "ccbg",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "cleanWin": "del /s /q dist",
+    "clean": "rm -rf dist",
+    "build": "tsc --build",
+    "watch": "tsx --watch"
+  },
+  "type": "module",
+  "keywords": [],
+  "author": "HirziDevs",
+  "license": "MIT",
+  "dependencies": {
+    "@minecraft/server": "^1.12.0-beta.1.21.1-stable"
+  },
+  "devDependencies": {
+    "ts-node": "^10.9.2",
+    "tsx": "^4.15.7",
+    "typescript": "^5.5.2"
+  }
+}
