+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "target": "ES2023",
+    "module": "ESNext", // Use ES Modules
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "outDir": "scripts/dist"
+  }
+}
